backend/requirements.txt
# Core Web Framework
fastapi==0.110.1
uvicorn[standard]==0.25.0
email-validator==2.3.0

# Authentication & Security
bcrypt==4.1.3
python-jose==3.5.0
python-dotenv==1.2.1
python-multipart==0.0.22

# Database
pymongo==4.5.0
motor==3.3.1
firebase-admin==7.2.0

# Google AI / Generative Models
google-genai==1.61.0
google-ai-generativelanguage==0.6.15

# Cloud & Storage
cloudinary==1.44.1
boto3==1.42.42
botocore==1.42.42

# Image / Vision Processing
opencv-python-headless==4.13.0.90
pillow==12.1.0
numpy==2.4.2

# Rate limiting / utilities
slowapi==0.1.9
requests==2.32.5

# Optional / Helper
matplotlib==3.10.6
pandas==3.0.0